Pull up sortMetrics.
