+++ components/frontend/src/subject/SubjectDetails.js
@@ -1,6 +1,5 @@
-import React, { useContext, useState } from 'react';
+import React from 'react';
 import { Dropdown, Icon, Table } from 'semantic-ui-react';
-import { DataModel } from '../context/DataModel';
 import { Metric } from '../metric/Metric';
 import { get_metric_comment, get_metric_issue_ids, get_metric_name, get_metric_status, get_metric_tags, get_metric_target, get_metric_value, get_source_name } from '../utils';
 import { ColumnMenuItem, HamburgerMenu } from '../widgets/HamburgerMenu';
@@ -58,70 +57,12 @@ function SubjectTableHeader({ hiddenColumns, toggleHiddenColumn, sortColumn, sor
     )
 }
 
-function sortMetrics(datamodel, metrics, sortDirection, sortColumn) {
-    const status_order = { "": "0", target_not_met: "1", near_target_met: "2", debt_target_met: "3", target_met: "4" };
-    const sorters = {
-        name: (m1, m2) => {
-            const attribute1 = get_metric_name(m1[1], datamodel);
-            const attribute2 = get_metric_name(m2[1], datamodel);
-            return attribute1.localeCompare(attribute2)
-        },
-        measurement: (m1, m2) => {
-            const attribute1 = get_metric_value(m1[1]);
-            const attribute2 = get_metric_value(m2[1]);
-            return attribute1.localeCompare(attribute2)
-        },
-        target: (m1, m2) => {
-            const attribute1 = get_metric_target(m1[1]);
-            const attribute2 = get_metric_target(m2[1]);
-            return attribute1.localeCompare(attribute2)
-        },
-        comment: (m1, m2) => {
-            const attribute1 = get_metric_comment(m1[1]);
-            const attribute2 = get_metric_comment(m2[1]);
-            return attribute1.localeCompare(attribute2)
-        },
-        status: (m1, m2) => {
-            const attribute1 = status_order[get_metric_status(m1[1])];
-            const attribute2 = status_order[get_metric_status(m2[1])];
-            return attribute1.localeCompare(attribute2)
-        },
-        source: (m1, m2) => {
-            let m1_sources = Object.values(m1[1].sources).map((source) => get_source_name(source, datamodel));
-            m1_sources.sort();
-            let m2_sources = Object.values(m2[1].sources).map((source) => get_source_name(source, datamodel));
-            m2_sources.sort();
-            const attribute1 = m1_sources.length > 0 ? m1_sources[0] : '';
-            const attribute2 = m2_sources.length > 0 ? m2_sources[0] : '';
-            return attribute1.localeCompare(attribute2)
-        },
-        issues: (m1, m2) => {
-            let m1_issues = get_metric_issue_ids(m1[1]);
-            let m2_issues = get_metric_issue_ids(m2[1]);
-            const attribute1 = m1_issues.length > 0 ? m1_issues[0] : '';
-            const attribute2 = m2_issues.length > 0 ? m2_issues[0] : '';
-            return attribute1.localeCompare(attribute2)
-        },
-        tags: (m1, m2) => {
-            let m1_tags = get_metric_tags(m1[1]);
-            let m2_tags = get_metric_tags(m2[1]);
-            const attribute1 = m1_tags.length > 0 ? m1_tags[0] : '';
-            const attribute2 = m2_tags.length > 0 ? m2_tags[0] : '';
-            return attribute1.localeCompare(attribute2)
-        }
-    }
-    metrics.sort(sorters[sortColumn]);
-    if (sortDirection === 'descending') {
-        metrics.reverse()
-    }
-}
-
 export function SubjectDetails({
     report,
     reports,
     report_date,
     subject_uuid,
-    metrics,
+    metricEntries,
     changed_fields,
     handleSort,
     sortColumn,
@@ -133,12 +74,6 @@ export function SubjectDetails({
     extraHamburgerItems,
     reload
 }) {
-    const dataModel = useContext(DataModel)
-
-    let metricEntries = Object.entries(metrics);
-    if (sortColumn !== null) {
-        sortMetrics(dataModel, metricEntries, sortDirection, sortColumn);
-    }
     const subject = report.subjects[subject_uuid];
     const last_index = Object.entries(subject.metrics).length - 1;
